Updated openstack/openstack Project: openstack/cinder  6897f23ee4f596ff5687c75e3b5fba1e646eb29c
VMAX driver - failover error fix

Issuing the failover command when the backend is already
failed-over, or the failback command (denoted by specifying
"--backend_id default") when the backend is not failed over
does not stop immediately, and attempts to fulfill the
command in the incorrect state. This could result in
inconsistencies, and also leads to confusing error messages
for the end user. This patch rectifies the issue and logs
appropriate errors.

Change-Id: I58b2e030cd5a2e3a89a030da0643a61cb7264d64
Closes-Bug: #1660383