Updated openstack/openstack Project: openstack/cinder  4bf39cd36dffb2793be2f13a20780010033a4311
HNAS: Cloned volume with different volume type

HNAS NFS driver allows the creation of cloned volumes using volume
types that are not the same as the ones from the source volume. This
could make cinder information inconsistent with the real volume
location in the backend when we have multiple pools configured.
As the clones are always created in the same pool as the source
volumes, this patch makes the driver accept the creation of cloned
volumes only if they have the same volume type as the source volume.

Closes-Bug: #1660411

Change-Id: I1272a6d27d7f172e77f1260a2acc6fc7c02e2e73