Updated openstack/openstack Project: openstack/neutron  4bc5ab4f0044cc9edbaa88359d0d5e26ecfd5cf7
Use same session object for ml2 dict extend functions

This adjusts the ML2 dict extend functions to pass the
same session to the drivers that was used to generate
the sqlalchemy object that will be passed to them. This
will prevent reloading objects in a separate session that
can result in a performance hit or reads to an entirely
different server than the one that the object was generated
from.

Partially-Implements blueprint: enginefacade-switch
Change-Id: Id2dbb13e8aa5fcb86fab8e815033e57e6f0e98e8