Updated openstack/openstack Project: openstack-infra/project-config  0c0480e1d5047b32233e7f2bb34d42d319f711a9
OSA: Ensure trusty job inclusion is not too broad

Currently the trusty job inclusion negates the rule above
which restricts very specific jobs for the stable/mitaka
branch, resulting in jobs running that should not.

This patch focuses the trusty inclusion to stable/newton
only as the trusty inclusion for stable/mitaka is already
handled in the rule above.

Change-Id: I7f5430705cac31672fe1e47eec8fc240abed9dab