Updated openstack/openstack Project: openstack/heat  b2e436aa67b373c4424cce7813ed1a0b0e54a734
Just to migrate existing resource to backup stack

Don't create a resource record for backup stack if the
resource record exists, just to migrate it to backup
stack, to avoid redundant data remaining for existing
stack.

This patch also adds resource.store() which covers
_store() and _store_or_update() implemention. And then
we can delete the two methods.

Change-Id: I0b4b983306ea84fab0e2c81876ef407a80d25989
Closes-Bug: #1662095