Updated openstack/openstack Project: openstack/python-neutronclient  30fdde8e4e7687ee9a2e8e00f8c0630abbaf1c48
x-openstack-request-id logged twice in logs

In the recent release of keystoneauth1 2.18.0 provision is made to log
x-openstack-request-id for session client. Once this new library is synced
in openstack projects, the x-openstack-request-id will be logged twice
on the console if session client is used.

For example,
$ neutron --debug port-list

DEBUG: keystoneauth.session GET call to network for
http://10.232.48.204:9696/v2.0/ports.json used request id
req-da75468a-2855-4e59-b308-d2a91776b927

DEBUG: neutronclient.v2_0.client GET call to neutron for
http://10.232.48.204:9696/v2.0/ports.json used request id
req-da75468a-2855-4e59-b308-d2a91776b927

Above log will be logged twice on the console.

Removed logging of x-openstack-request-id in case of SessionClient as it
is already logged in keystoneauth1. x-openstack-request-id will only be
logged once on console if HTTPClient is used.

Depends-On: I63fb5e5486670bc369abc5339a2da8e65ba2ec6f
Closes-Bug: #1657351

Change-Id: I8efefd69a1ba4a214b1a30b20b9c1e5bd96b19d4