Updated openstack/openstack Project: openstack-infra/shade  5343ade6a7e5e5f7f5f467314cbb9d913c705f13
Transition nova flavor tests to requests_mock

We had added a betamax fixture for create_flavor, but the requests_mock
approach is actually working out much better. Go ahead and replace it to
simplify the test suite a little bit.

Remove _by_flavor tests as they tested a behavior that's actually invalid but
worked in the test by happenstance.

Transition the rest of the file while we're in there.

Change-Id: Ic2457d7380a8af41ed7bf6b264cbdc2240780ff3