Updated openstack/openstack Project: openstack/nova  06e400384c432ebbefe75029ba3ae7791f367c40
Update the upgrades part of devref

This removes some references to "online schema migrations", as well
as some references to in-progress things that have been (long since)
completed. It also clarifies some of the upgrade steps, and unifies
the notion of "offline" and "live" upgrades, calling out only a couple
places where the process differs.

This came from me explaining the document to someone and calling out
things that were no longer accurate.

Change-Id: Iec1d73c3fd0038d9bca7e7fa746caad06c329ed3