Updated openstack/openstack Project: openstack-infra/shade  f2f608529502dacbfdb44880de5be27cf85596d5
Port in log-on-failure code from zuul v3

Zuul v3 picked up a great technique for logging a ton of useful
information for errors but also not blowing out the subunit stream on
success.

We do it a little differently here - we don't have the need for the
flexibility in logging config that zuul has. Also, add the ability to
set an env var and print the logging even on success. (useful when
developing single tests when one wants to inspect that a success really
did what we think.

Change-Id: Ibcb9a4aa33c16450df252c5e0f32a62c5cc821bc