Updated openstack/openstack Project: openstack/cinder  0cd5244686ad644ae4e68851c31f1f2ac520b0f9
xiv create vol from replicated source fails

The ibm_storage driver fails to clone a volume from a replicated
source. It creates the volume, defines replication for it,
and when it tries to copy the contents of the source volume,
it fails. The logic should be reversed - after create, we should
copy the contents of the source and then set the replication

Change-Id: Ic6d093bfc3c03807211dd47c5eb4428127e28821
closes-bug: 1661223