Updated openstack/openstack Project: openstack/neutron  866b184d93086ff5ef1b0ab2043651a07e2a3696
DVR: Add static routes to FIP namespace

The static routes are currently only added to the router
namespace and snat namespace, but not added to the fip
namespace.

This patch adds the static routes configured for the router
to the FIP namespace, to its own table. So this will allow
the FIP namespace to be configured with different nexthop
routes for different routers.

Change-Id: Ida165d1ecf5c07af31dac11d9daed33ccaaf5605
Closes-Bug: #1571676