Updated openstack/openstack Project: openstack/oslo.log  9c20b21cb282769fe65eaaf202ec5854a372cec2
Remove 'verbose' option (again)

The deprecation for this was merged on aug 1, 2015 in review:
https://review.openstack.org/#/c/206437/
That was a long time ago so it should be possible to remove it now.

This was already merged once but had to be reverted because some
projects were still relying on it. Let's try again, this time using
INFO default not WARNING, thanks dtantsur!

Change-Id: Iac62cb252db0e8d86c352210089d70405edb5c4d