Updated openstack/openstack Project: openstack/oslo.versionedobjects  e484ffbe1eac108b09393808c9e0de62afba948e
fields: fix stringifying of objects when subclasses=True

When the 'Object' field stringifys a value, it is
currently hardcoded to report the object name that
the field was initialized with. This used to be
sufficient, until the 'subclasses' parameter was
added. When subclasses=True, the value may be any
arbitrary subclass, so we should be reporting the
actual object name associated with the value.

Closes-Bug: #1638577
Change-Id: I0bb4ade30a60554710a12f0f058945ef43826eff