Updated openstack/openstack Project: openstack/diskimage-builder  cac1471cc82e635e892ed8a0e580babf295c74c4
yum/install-packages output cleanup

Rather than echo the package list separately, turn on tracing around
the yum/dnf call so that we get more complete output.  No loss of
info, as this effecitvely shows the packgae list; plus it has the
bonus that it makes more sense in the logs, because otherwise you get
all the yum/dnf output without knowing what the call was.

Change-Id: I44df6e944af602e6b03f669c15a521056de1ad79