Updated openstack/openstack Project: openstack/oslo.db  02247b3a0ec75d96acfff740b935e07cbac32a1b
Explain paginate_query doesn't provide parameter offset

OFFSET is a terrible thing to use.   It seems to work great when you
first use it because your table is only a few hundred rows.  six months
later, everyone complains the app is incredibly slow - it's because the
database is being killed doing linear scans through millions of rows
because everyone's doing OFFSET.  It not only causes slow and unusable
applications but hurts everyone by adding tons of meaningless work to
the DB load overall.

This pagination function already does the "right" way, which is to order
the rows and use WHERE criteria  (See http://use-the-index-luke.com/no-offset).

This commit add docstring to explain why don't need offset.

Co-Authored-By: Mike Bayer <mike_mp@zzzcomputing.com>
Change-Id: I6fdb094085d133c8910415efea2eb44225a66216