Updated openstack/openstack Project: openstack/oslo.vmware  6e4d16d9113d896fd6061da6a3e54275b396e0aa
Add in support for removeKey

When setting a removeKey the VC would return an error indicating that it
"Unable to determine type of Any". We need to update this to ensure that
the type set is a string or an integer. This is due to the fact that it
can be both.

Change-Id: I494a1d19341f30f22803a3fe6baf020a67ad6b08