Updated openstack/openstack Project: openstack/oslo.policy  9e3d46b1707152094cc2c2bdd45e22898d79140c
Add optional exception for check_rules

Rather than checking for a True/False return from check_rules some
consumers may want it to raise an exception instead. This adds a
parameter, defaulting to False, which will cause check_rules to raise an
exception if set to True and a rule violation is found.

Co-Authored-By: ChangBo Guo(gcb) <eric.guo@easystack.cn>
Change-Id: Ie69bbc7081e84d019b85b62a45589ed331cc1f15