Updated openstack/openstack Project: openstack-infra/jeepyb  1c7264f9252232574e6498c12fe96a0d18fad062
Cache status of stages in manage-projects

Implement a trivial caching system that will allow us to skip even
attempting gerrit and github remote operations for projects that we've
processed successfully in previous passes. This should hopefully reduce
the amount of time we spend processing the projects.yaml file.

In the future, we could extend this to tracking more specific actions-
such as the description set in github, so that we could know whether or
not they have changed and thus need to be processed again.

In this form, acls files will always be processed.

Change-Id: I07b13c8663e6f9ee1255a4e56caf556ea49fb51b