Updated openstack/openstack Project: openstack-infra/jeepyb  3b7adef6fb90f45d480a121ad9e7ace94d1feff4
Check git fsck output when importing projects

This is a sanity check to properly log when repos are corrupt and not
able to be imported into Gerrit. The import will fail either way but at
least with this change in place we will get more direct logging of the
reason for the import failure.

Change-Id: Ibb1839329065e5d4a9747f95ff21029148801874