Updated openstack/openstack Project: openstack/cinder  aec465c0b2d740c59b6cc773a9c9c469c3bc4b94
VMAX driver - copy state fix

Creating a cloned volume can fail when the source volume
is bigger than 50 GB. This occurs when we try to unlink the
source from target before all the tracks are copied.
This patch fixes the issue by adding a 'WaitForCopyState'
parameter to the CreateElementReplica SMI-S function.

Change-Id: Iac8ce48dda6f21a49838e4ca900a4f3ba1a9b2a7
Closes-Bug: #1660378