Updated openstack/openstack Project: openstack/manila  3726c89f20d48929a498ad61d1270014cef8a223
Pass access rules to driver on snapshot revert

In order to revert to a snapshot in the LVM driver (and
very likely other drivers) the list of access rules is
needed, so this change modifies the driver interface to
provide this extra information.

This change requires preventing a revert to snapshot
operation while access rules on the affected share are
out of sync.

Closes bug: 1658133

Change-Id: Ia6678bb0e484f9c8f8b05d90e514801ae9baa94b