Updated openstack/openstack Project: openstack/ironic  9da91aaaf5d8f5f9b97497615d5fe4ad74c9e59e
log if 'flat' interface and no cleaning network

Commit cde11611d93946a1c79e406dc429aa5e742da729 added a TODO (in
Ocata) to raise DriverLoadError if using 'flat' network interface
and cleaning_network config option is not set.

After discussions, we decided to keep the logging behaviour,
because one could use a flat network and not do any cleaning.
The logging message is updated to reflect this.

Change-Id: Ida3f3a349655bffc3ed0a307a3c18fbd69b42271
Fixes-Bug: 1661082