Updated openstack/openstack Project: openstack-infra/shade  93c2e17c8ec1280fec4473e9f583cb2b29c03073
pass -1 for boot_index of non-boot volumes

Although the API documentation for nova [1] says to pass None to
indicate a volume that should not be used for booting, doing so produces
an error because None is not an integer. Change shade to use '-1' (to
comply with the declared string type) instead of None.

[1] http://developer.openstack.org/api-ref/compute/?expanded=create-server-detail#create-server

Change-Id: I5732575501005f4acc6320578882a079adc8be38
Signed-off-by: Doug Hellmann <doug@doughellmann.com>