Updated openstack/openstack Project: openstack/tripleo-heat-templates  d5b44ae7dc03bf999b696f540061220fe6137b11
Stop deploying Nova API in WSGI with Apache

It was suggested by Nova team to not deploying Nova API in WSGI with
Apache in production.
It's causing some issues that we didn't catch until now (see in the bug
report). Until we figure out what was wrong, let's disable it so we can
move forward in the upgrade process.

Change-Id: I09b73476762593642a0e011f83f0233de68f2c33
Related-Bug: 1661360