Updated openstack/openstack Project: openstack-infra/project-config  6842e232a6fb3def372fddaced481df7e430e193
Create ansible-role-fedmsg project

Fedmsg is a brokerless messaging bus created by the Fedora project.
Recently we added support to zuulv3 to report results to it.  As such,
create an ansible role we can use to properly manage fedmsg.

This project is not currently part of the big tent, but will be used
under the existing windmill ansible roles.  We are also adding IRC
notifications into the #windmill channel.

Change-Id: I646e93a19216574d5750a06ed2138df4ed3b7b00
Signed-off-by: Paul Belanger <pabelanger@redhat.com>