Updated openstack/openstack Project: openstack/neutron-fwaas  86d46aae8191ad39fdca91c1e3914d0b4b27a33e
Do not complain in firewall_deleted if the FW is already deleted

Currently firewall_deleted[1] crashs if the firewall is already deleted
or deleted concurrently during firewall_deleted call. We should avoid
such behavior as there is no reason to crash if someone already did the
job for us (ie: delete the FW).

Moreover such crash is costly because it triggers a service-sync on FWaaS
l3-reference agent (at least). Typically on a L3-DVR deployment, all
firewall_deleted calls except the first one will fail so quite every
L3-DVR will perform a FWaaS service-sync.

This change updates firewall_deleted in order to succeed if the firewall
is already deleted or if the firewall is deleted concurrently.

Such trouble also impacts FWaaS v2 and will be solved in a follow-up
change.

[1] neutron.services.firewall.fwaas_plugin.FirewallCallbacks

Partial-Bug: #1658060
Change-Id: I4df05f2026eb622662dad82e504d072639cdae48