Updated openstack/openstack Project: openstack/nova  50d402821be7476eb58ccd791c50d8ed801e85eb
Consider startup scenario in _get_compute_nodes_in_db

Before this change, on the first start of the nova-compute
service on any host, this method is logging the
"No compute node record for host" error which is confusing
for people debugging issues. For example, if the compute
service starts up before the Placement service is running,
the compute is checking in and failing to connect to the
placement endpoint which might take the compute node out
of consideration for scheduling. When debugging that kind
of issue, people can get hung up on this error message
which is actually an expected case on the first start of
nova-compute on a new host.

This change simply plumbs through a boolean to tell it
if we're starting up or not. This doesn't tell us if
it's the first time we're starting this service or a
restart, but the thinking is, if it's a restart we don't
get the NotFound error and if we do, then we do want the
error level message.

Change-Id: Id7a05b579ead6ac5445ca5b1eeab6d223d545a6c