Updated openstack/openstack Project: openstack/horizon  6fa0a90318d8d299587c3e871767163dc1556b2b
Make shared image text less confusing for Glance v2

When using Glance v2 and logged in as an admin, the images
panel now shows all the images in the cloud. This is the
way the Glance v2 list api works, but it changed the behavior
from v1. In Horizon, we can't tell whether non-public images
that aren't owned by current project are shared or just from
some other project without making multiple api calls. This
patch makes the text of the images less confusing when using
Glance v2, so that it no longer claims the images are "Shared
with Project".

Change-Id: I2859e104de78a6a633b0e1a2ff30dde674b4bdee
Closes-Bug: #1624743