Updated openstack/openstack Project: openstack/nova  e7542451de834bfde90349ae0b7a7ec11eb41d1f
Add an update_cell command to nova-manage

In case a cell record was incorrectly created, or needs to be updated
to point to a new database or mq endpoint, we need a way for users to
be able to update a CellMapping record. Since deleting and re-creating
is definitely not an option, this patch adds an update_cell which
allows updating the mutable fields if necessary.

Change-Id: I75ddfa458a6fb650f1098505193398984f2be596