Updated openstack/openstack Project: openstack/neutron-fwaas  dc27934dc4bca938002616ed753c8c69d2453037
Do not complain in firewall_group_deleted if the FW is already deleted

Currently firewall_group_deleted[1] crashs if the firewall is already
deleted or deleted concurrently during firewall_deleted call. We should
avoid such behavior as there is no reason to crash if someone already
did the job for us (ie: delete the FW).

Moreover such crash is costly because it triggers a service-sync on FWaaS
l3-reference agent (at least). Typically on a L3-DVR deployment, all
firewall_deleted calls except the first one will fail so quite every
L3-DVR will perform a FWaaS service-sync.

This change updates firewall_group_deleted in order to succeed if the
firewall is already deleted or if the firewall is deleted concurrently.

[1] neutron.services.firewall.fwaas_plugin_v2.FirewallCallbacks

Change-Id: Ic0b228896c8129205224417506bb06471e432955
Closes-Bug: #1658060