Updated openstack/openstack Project: openstack-infra/tripleo-ci  2f7e6b97370e6efdf69ca41a67cbf773fcb41fec
Pass job identifier to testenv-worker

Adds another argument to the call to the te-broker that will allow
us to identify which patch and job a given testenv was used for.

The arguments passed in by the client already get logged on the
te-broker side, which is why there are no changes except in the
client.

Change-Id: I848e08cd272724c059ce92a6ee527dedcd53acce