Updated openstack/openstack Project: openstack/nova  9e555dc5c178f659c3d49784dc1568714f9393df
Release PCI devices on drop_move_claim()

On cold migration, drop_move_claim() is called in the confirm stage on the
source node.  Since the migration is being tracked by the resource tracker on
the destination node, the source node has the instance in it's
tracked_instances.

So in this case the PCI devices were only freed on the next periodic audit.
For PCI resources such as PCI passthrough, those are limited in number and
should be freed right away.

This patch fixes drop_move_claim() to also free PCI devices when an instance
is in self.tracked_instances().

Co-Authored-By: Steven Webster <steven.webster@windriver.com>
Change-Id: Ie3392f80dfd2650048519c571ffaa11c025ad048
Closes-Bug: #1641750