Updated openstack/openstack Project: openstack/nova  c42a7e172b51e9c36445bb724beeb1e9a71721f8
Assign mac address to vf netdevice when using macvtap port

This is a workaround to libvirt issue described here:
https://bugzilla.redhat.com/show_bug.cgi?id=1372944. The problem is that
libvirt dosn't update the mac address of the netdevice, instead it uses
netlink to configure the mac address of the VF. For macvtap to work
libvirt should set the mac address on the VF net device and using
netlink. Until this is fixed we should set both of them in nova.

Change-Id: If32828840b552a73400df052fc12a8b5a24f202d
Closes-Bug: #1622292