Updated openstack/openstack Project: openstack-infra/zuul  10b71820c41c300c1959d83af3d42fa4e0ee5ab4
Set keepalives for gerrit connections

Some gerrit clients connect through stateful firewalls,
which are occasionally "flushed", which leads to hanging of the connection.
This could be avoided by tcp keepalives.
To turn keepalives on, set "keepalive" in gerrit connection attributes to non-zero
(0 means no keepalives sent)
The default is turn it on with timeout of 60 seconds.

Change-Id: Ic3d11e0ac2108f02832e2114ba8bc1adcc87cb4f