Updated openstack/openstack Project: openstack-infra/shade  1a1d1a41c4d3c760a0727a566e4b32d43cf35be8
Transition half of test_floating_ip_neutron to requests_mock

There is too much going on here for a single patch, and my brain hurts
too.

We need to come back through and refactor test_auto_ip_pool_no_reuse.
The resources involved are all accurate (took them from a live
interaction with citycloud for verification purposes) but doing complex
tests that way is a pile of copy-pasta.

Also, found a logic error in the flow of having created a server and
then separately requested a FIP be attached to it with reuse=False. The
logic error is that the check to see if the server has a FIP already is
using a stale copy of the server dict which cannot possibly have a FIP
attached.

Change-Id: Ia38294cf48bcb46c18892605fd52da7dadaeefc8