Updated openstack/openstack Project: openstack-infra/tripleo-ci  92fdff1ae9d320299f2f71300d578b7b063397e4
Use only RAMWeigher for rh1 scheduling

We have a few black sheep compute nodes in rh1 that have no SSD
and/or less memory.  The non-SSD nodes tend to get preferred by the
scheduler due to the larger disk number they report (1000 GB > 200 GB)
and that's particularly a problem when we have a node with only 64
GB of memory and a 1 TB drive.  It tends to get over-scheduled,
even though it's our slowest node.

Since we almost exclusively care about distributing memory evenly,
let's weigh scheduler decisions only on that.  Note that the normal
filters are left in place, so we shouldn't ever try to schedule
more vms than a node can handle.  This will only stop us from
preferring the nodes with slower storage.

Also note that this change is already live in rh1 and seems to be
working fine.  I'm just updating the env to reflect the change.

Change-Id: I54731e49ed9bb08a6048bc52fe25412d9de6473c