Updated openstack/openstack Project: openstack/manila  65754aa1539d4acfadb827b6986a0a68f22b7b9a
HNAS: ensure snapshot before trying to revert

HNAS driver does not need to start reverting if ensure_snapshot fails,
it prevents trying to revert when snapshot does not exist on backend
anymore.
Also, an additional check for snapshot path in backend was added in
ensure_snapshot. It is needed to ensure path for shares without
'mount_snapshot_support'.

Closes-bug: #1662569
Change-Id: Idfb02193f012017cdc87468ee221bfa8f22c3fd2