Updated openstack/openstack Project: openstack/requirements  9d723340207226f4fe5e0ab79466c7fa9a410908
Update sphinx requirement

In change I38580a5f765c7bb586dcf2786e91be15a8eba235 we skipped one of the
docutils versions due to a breaking release that was not supported by
sphinx. This has now been fixed in the latter[0].

With Sphinx 1.5.1, the blocking of docutils 0.13.1 is not needed
anymore, so remove the condition. This line could be removed completely,
but for that it first needs to be removed from all projects.

[0] https://github.com/sphinx-doc/sphinx/issues/3212

Change-Id: I01b7b6f3fac75e284debe211129b7b1e87c756d9