Updated openstack/openstack Project: openstack-infra/project-config  a46ed6c730359441049124d853807cf54734c713
Rename placement job to nova-next

This job is really more than just testing nova with
the placement service (which was new in newton), since
it's also running with cells v2 (also optionial in newton).

The intent of this job is to test the latest features in nova
which for Pike would be things like cinder v3 and
sending service user tokens with regular user tokens (both of
which are available in pike).

We can control the features enabled in the builder based on
the branch the job is running against.

As of this patch, this job is still only running in stable/newton.
Once we start adding some more features for testing in Pike, we
can make this run in master (Pike) also.

Change-Id: Idc3a3c82eaaedb9bd75f9bb4774469853d453899