Updated openstack/openstack Project: openstack/horizon  b7730aba0dd88dcc075d4496504864d9b2c488e5
Use "Default", "Value" and "DefaultValue" for Heat parameters

It seems there has been changes in the Heat API, and it no longer
returns the default value for a parameter in the "Default" key, it
has been instead renamed to "DefaultValue" and later to "Value".
This patch makes it use all those key names, whichever exists.

Change-Id: Ia86d8d79fcd48974ffe3fecf21a7153bf2ba6a9a