Updated openstack/openstack Project: openstack-dev/devstack  e0a37cf21e43fbb4ba3f9f8fa5321a0a0e1bedf1
tls proxy: immediately close a connection to the backend

Force mod_proxy to immediately close a connection to the backend
after being used, and thus, disable its persistent connection and
pool for that backend.

Let's see if that helps fixing bug #1630664 (the
Connection aborted/ BadStatusLine thing).

We already have an ER query (in queries/1630664.yaml) that should show
whether this is effective.

Change-Id: I03b09f7df5c6e134ec4091a2f8dfe8ef614d1951