Updated openstack/openstack Project: openstack/tempest  88d84e951b8bea60d7a65cb8a9449529b1716a01
test_create_server_from_volume_snapshot: assert dict is not empty before accessing a key

Following the patch: If2a257b24f2d50a401d6804e983bcad55b5c89e8
In case the dict is empty i.e. there in not an attachment, it will
be helpful to check whether there is an attachment or not to prevent
error in L204.

Change-Id: Ib74472b3b147f34a63ac5988c7db97230a9f4803