Updated openstack/openstack Project: openstack/requirements  3bf3bb13337e2b85895d3aeea87167d9e12fa8d6
Add python-vitrageclient to the global-requirements.

python-vitrageclient is an actively developed OpenStack project, and it is consumed by vitrage-dashboard project.

Answers to the requirements Q&A:

Q: Is the library actively maintained?
A: Yes, by the Vitrage team

Q: Is the library good code?
A: We are doing our best :-) working by OpenStack standards

Q: Is the library python 3 compatible?
A: Yes

Q: Is the library license compatible?
A: Yes, it uses Apache-2.0 like the rest of OpenStack

Q: Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
A: It is packaged in Fedora (https://apps.fedoraproject.org/packages/python-vitrageclient/overview/)

Q: Is the function of this library already covered by other libraries in global-requirements.txt?
A: No

Q: Is the library required for OpenStack project or related dev or infrastructure setup? Which?
Yes. It is used by vitrage-dashboard

Change-Id: I4cb2f09c8fe4ca2d3edde49b6083686f1a6f850b