Updated openstack/openstack Project: openstack/requirements  ef04080b0032ed9be9de02a534f319c59b5fa405
remove upper cap of kombu/amqp

kumbu/amqp upper caps have been set because their API have changed in
4.0.0/2.0.0

We are working on oslo.messaging side to use the new API [1]. But we
can't test it in gate due to the current version in global requirements
and upper-constraint.

This change removes the upper cap to allow oslo.messaging update.

[1] I73958c8057353a2eefe1baaa7a41148193d507f7

Change-Id: I9f980b51901ac31599b9651633956ad2eea6a1ac