Updated openstack/openstack Project: openstack/neutron  521b1074f17574a5234843bce68f3810995e0e1d
Pecan: Fix tags handling for pecan

Tags controller was not being picked up from the extension.  This also
allows the usage of collection_methods that tags uses.  Also, the tags
extension made use of action_status on a resource, which allowed
overriding the default status codes, so this patch handles action_status
as well.

Change-Id: Ic7f5cfddd30f26737469516235053656f1ba6a24