Updated openstack/openstack Project: openstack-infra/tripleo-ci  ae722fa9a86af6eb39e13889481d3fdcad58e65e
Set correct IP in node and node_private

When using --setup-nodepool-files for local CI multinode testing, the
/etc/nodepool/node and /etc/nodepool/node_private files should be
created for consistency with the files that are expected to exist.

We are also making use of these files in our networking setup for
multinode, so they should exist when trying to replicate those
envirionments.

Change-Id: Iebf0e855ecddcd9d544a9e264cafd6c2698bc8dd