Updated openstack/openstack Project: openstack-infra/project-config  906683f47eb5bcade800c6cc46658659f8854b1f
Move OpenStack-Ansible os_tempest Centos to non-voting

CentOS gates in OSA are taking considerably longer than Xenial gates.
This is causing a large number of timeouts, and is blocking the gate
for roles that have longer gate runs.

We should move the os_tempest role back to non-voting until a more
permanent solution can be achieved.

Change-Id: I2b8b7f2eabd337eb04b5d890449f243012e7b114