Updated openstack/openstack Project: openstack/oslo.rootwrap  1f60ecbb9d7b0c39df9d7d663173ea892f4d4378
Always check cmd which does not exist

In daemon mode, if run a cmd which does not exist,
it will raise an Exception. But when we install the
cmd later and run cmd, it still raise an Exception.
The only work around right now is to restart the
daemon, with this change, we can avoid restarting
the daemon

Closes-Bug: #1663216
Change-Id: I6f6ff540ed103f0fec329d6264bdac26493c8fad