Updated openstack/openstack Project: openstack/tempest  bad7dcd77c5954e12c871e7d62300ff42addfbeb
Remove skip of test_stamp_pattern

test_stamp_pattern is skipped of bug 1205344 which is reported at 2013,
though there is no exact hint on whether the bug is fixed or not, but
since 2013 is too long ago, we guess the bug should have been fixed.

This is to remove the skip, and to fix several obvious errors in the
code.

Change-Id: I2c13099595d8ee8099e02e3f06096078c7d27c34