Updated openstack/openstack Project: openstack-infra/irc-meetings  bab254438aeab9ced07ce256ac436dcef5abebfc
Remove networking-ovn meeting.

We have not been using this timeslot recently.  Free it up for someone
else to take.  If we need to discuss networking-ovn outside of the
main OVN meeting, we need to pick a different time slot anyway, since
after DST, this now overlaps with the main OVN meeting.

Change-Id: I7d79feebdc71e9057f06d647594181ac4b715425