Updated openstack/openstack Project: openstack/neutron  6759af1fa95499bbda7634c8e801a496c199b38c
Add helper for driver to check if it owns router

This adds a helper that an L3 service provider can use to check
if a given router_id is associated with it. This makes skipping
events unrelated to the driver simple calls. See dependent patches
for examples.

Related: blueprint multi-l3-backends
Change-Id: I3ecb935722dd7fe75f73ad410225ff1b93b0f605