Updated openstack/openstack Project: openstack-infra/jeepyb  6d1ba2f1696a9f8f502e39b804505cfa2a690e22
Continue to optimize manage-projects for resiliency

Saving the state for having created the project was sticking to false.
If we had the git repo locally but it wasn't in gerrit, the logic was
all message up. Checking for homepage and downloads in github kills us -
although it would be fine if we'd started from caching homepage and
description in the first place, so maybe we need a script that sets
status quo in the cache file.

Change-Id: I15802082c0bef9e9362eb9482b058f81d1e566d7